Moved AuthTokens.java to its own folder 'auth'.
